requirements.txt
plotly
streamlit
azure-mgmt-network
azure-mgmt-compute
azure-mgmt-web
boto3
botocore
python-dotenv
Pillow
openai
azure.identity
azure.mgmt.resource
matplotlib
easyocr
#cv2
requests
fastapi 
uvicorn 
python-multipart
pulumi>=3.0.0
pulumi-automation>=0.1.0
pulumi-azure-native>=1.0.0
azure-identity>=1.0.0

